feat(ui): added an alert to conforim deletion of posts and comments. Comments can now be deleted
Added a static method in UiUtils for confirmation. Deletion of comments has also been implemented.
Since we have refactored anonymous author, the user can now also delete anonymous posts/comments.

Closes #78 #85
Development-method: individual programming

+++ messeption/ui/src/main/java/messeption/ui/FrontPageController.java
@@ -263,8 +263,13 @@ public class FrontPageController {
       if (deleteButton != null) {
         deleteButton.setOnAction(e -> {
           try {
-            boardAccess.deletePost(post.getId());
-            drawPosts();
+            boolean confirmation = UiUtils.confimationAlert("Confirm deletion",
+                "Delete post: " + post.getTitle(), "Are you sure you want to delete your post?");
+
+            if (confirmation) {
+              boardAccess.deletePost(post.getId());
+              drawPosts();
+            }
           } catch (Exception e1) {
             UiUtils.exceptionAlert(e1);
           }
